aaaaaaaaa

+++ Project blight/Detour.hpp
@@ -9,7 +9,7 @@ uintptr_t Detour_getCurrentSwingDuration(Player*);
 
 int Detour_getItemHanded(uintptr_t*, uintptr_t*, uintptr_t, BYTE);
 
-uintptr_t Detour_Update(__int64, __int64, __int64);
+uintptr_t Detour_Update(uintptr_t, uintptr_t, uintptr_t);
 
 uintptr_t* Detour_ConnectionRequest(uintptr_t*, __int64, __int64, __int64, __int64, __int64, __int64, __int64, __int64, __int64, __int64, int, int, int, __int64, char, char, __int64, int, __int64, __int64, char, __int64, __int64, __int64, char);
